Add search in platform version and ICD suffix
The platform selector from command line is now searched for in the
platform version, and is compared for a full match with the ICD suffix.

Add cl_khr_icd and cl_khr_fp64 to pre-defined OpenCL 1.2 extensions.

Platform and device names (selectors) from the command line are now
case-insensitive.

Fixed working directory for CMake custom command to generate tags file.